move hooks folder to root
